moved CourseRepository & fixed syntax error from previous merge
