+++ LMS.API/Extensions/ServiceExtensions.cs
@@ -94,7 +94,7 @@ public static class ServiceExtensions
         services.AddScoped(provider => new Lazy<IAuthService>(() => provider.GetRequiredService<IAuthService>()));
         services.AddScoped<IUserService, UserService>();
         services.AddScoped(provider => new Lazy<IUserService>(() => provider.GetRequiredService<IUserService>()));
-        services.AddScoped<ICourseService, CourseService>(); 
-        services.AddScoped(provider => new Lazy<ICourseService>()) => provider.GetRequiredService<ICourseService>()));
+        services.AddScoped<ICourseService, CourseService>();
+        services.AddScoped(provider => new Lazy<ICourseService>(() => provider.GetRequiredService<ICourseService>()));
     }
 }
